+++ buildSrc/src/main/kotlin/internal/VersionCatalogExtensions.kt
@@ -1,0 +1,8 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+import org.gradle.accessors.dm.LibrariesForLibs
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.the
+
+val Project.libs: LibrariesForLibs
+    get() = the()