+++ suspend-test/src/nativeMain/kotlin/SuspendTest.kt
@@ -1,0 +1,13 @@
+package com.javiersc.runBlocking
+
+import kotlin.coroutines.CoroutineContext
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.runBlocking
+
+/** Native `runBlocking` implementation */
+public actual fun <T> suspendTest(
+    context: CoroutineContext,
+    block: suspend CoroutineScope.() -> T
+) {
+    runBlocking(context, block)
+}